♻️ 0JV1JB context: decompose harvest artifacts
