+++ .agentplane/tasks/202605282313-0JV1JB/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282313-0JV1JB`
+Title: Harvest task artifacts decomposition
+Canonical task record: `.agentplane/tasks/202605282313-0JV1JB/README.md`
+
+## Summary
+
+Harvest task artifacts decomposition
+
+Decompose packages/agentplane/src/context/harvest-tasks-artifacts.ts by extracting stable model/build/render helpers while preserving the existing context harvest tasks API and behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/context/harvest-tasks-artifacts.ts by extracting stable model/build/render helpers while preserving the existing context harvest tasks API and behavior.
+- Out of scope: unrelated refactors not required for "Harvest task artifacts decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:13:46.861Z
+- Branch: task/202605282313-0JV1JB/harvest-artifacts-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
