+++ .agentplane/tasks/202605282313-0JV1JB/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 0JV1JB task: Harvest task artifacts decomposition [202605282313-0JV1JB]
