+++ .agentplane/tasks/202605282313-0JV1JB/pr/meta.json
@@ -1,0 +1,13 @@
+{
+  "base": "main",
+  "branch": "task/202605282313-0JV1JB/harvest-artifacts-decomposition",
+  "created_at": "2026-05-28T23:13:46.861Z",
+  "diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "last_verified_at": null,
+  "schema_version": 1,
+  "task_id": "202605282313-0JV1JB",
+  "updated_at": "2026-05-28T23:13:46.861Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
