+++ .agentplane/tasks/202605282313-0JV1JB/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T23:13:46.861Z
+
+## Task
+
+- Task: `202605282313-0JV1JB`
+- Title: Harvest task artifacts decomposition
+- Status: DOING
+- Branch: `task/202605282313-0JV1JB/harvest-artifacts-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282313-0JV1JB/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:13:46.861Z
+- Branch: task/202605282313-0JV1JB/harvest-artifacts-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
